# Invoice-System-Pro/lib/api-zod/src/generated/types/createStaffRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandro Industrial Invoice System API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffRole } from "./staffRole";

export interface CreateStaffRequest {
  name: string;
  role: StaffRole;
  active?: boolean;
}